+++ apps/router-demo/router-host-2000/cypress/e2e/remote2.cy.ts
@@ -1,13 +1,11 @@
-import { getH2, getH3, wait2s } from '../support/app.po';
-
 describe('router-remote2-2002/', () => {
   beforeEach(() => cy.visit('http://localhost:2002/'));
 
   describe('visit', () => {
     it('jump to home page', () => {
-      getH2().contains('Remote2 home page');
-      cy.get('.self-remote2-detail-link').click();
-      getH2().contains('Remote2 detail page');
+      cy.verifyContent('Remote2 home page');
+      cy.clickByClass('.self-remote2-detail-link');
+      cy.verifyContent('Remote2 detail page');
     });
   });
 });
@@ -17,12 +15,11 @@ describe('router-remote2-2002 in host', () => {
 
   describe('Remote2 render and destroy', () => {
     it('jump to remote2 home page', () => {
-      cy.get('.host-menu > li:nth-child(4)').click();
-      cy.get('.menu-remote2-home-link').click();
-
-      getH2().contains('Remote2 home page');
-      cy.get('.menu-remote2-detail-link').click();
-      getH2().contains('Remote2 detail page');
+      cy.clickMenuItem('Remote2');
+      cy.clickByClass('.menu-remote2-home-link');
+      cy.verifyContent('Remote2 home page');
+      cy.clickByClass('.menu-remote2-detail-link');
+      cy.verifyContent('Remote2 detail page');
     });
   });
 });