+++ apps/router-demo/router-host-2000/src/runtime-plugin/retry.ts
@@ -3,13 +3,13 @@ import { RetryPlugin } from '@module-federation/retry-plugin';
 const retryPlugin = () =>
   RetryPlugin({
     fetch: {
-      url: 'http://localhost:2001/mf-manifest.json',
-      fallback: () => 'http://localhost:2001/mf-manifest.json',
+      // url: 'http://localhost:2001/mf-manifest.json',
+      // fallback: () => 'http://localhost:2001/mf-manifest.json',
     },
     script: {
       retryTimes: 3,
       retryDelay: 1000,
-      moduleName: ['remote1'],
+      // moduleName: ['remote1'],
       cb: (resolve, error) => {
         if (error) {
           throw new Error(