+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @module-federation/cli
 
+## 0.12.0
+
+### Patch Changes
+
+- Updated dependencies [ebef2d0]
+- Updated dependencies [c399b9a]
+- Updated dependencies [ef96c4d]
+- Updated dependencies [f4fb242]
+  - @module-federation/dts-plugin@0.12.0
+  - @module-federation/sdk@0.12.0
+
 ## 0.11.4
 
 ### Patch Changes